Initial stab at separating the responsibilities of the Dangerfile
