+++ lib/danger/danger_core/dangerfile.rb
@@ -1,40 +1,39 @@
 # So much was ripped direct from CocoaPods-Core - thanks!
 
-require "danger/danger_core/dangerfile_dsl"
 require "danger/danger_core/standard_error"
-
-require "danger/danger_core/plugins/dangerfile_messaging_plugin"
-require "danger/danger_core/plugins/dangerfile_import_plugin"
-require "danger/danger_core/plugins/dangerfile_git_plugin"
-require "danger/danger_core/plugins/dangerfile_github_plugin"
 
 module Danger
   class Dangerfile
-    include Danger::Dangerfile::DSL
-
-    attr_accessor :env, :verbose, :plugins, :ui
+    attr_accessor :env, :verbose, :ui
 
     # @return [Pathname] the path where the Dangerfile was loaded from. It is nil
     #         if the Dangerfile was generated programmatically.
     #
     attr_accessor :defined_in_file
 
+    def initialize(env_manager, cork_board)
+      @ui = cork_board
+      @env = env_manager
+
+      env_manager.setup_plugins(self)
+    end
+
+    def extend_with_plugins(plugin_host)
+      # Create attributes for all of the external plugins
+      plugin_host.external_plugins.each do |plugin|
+        name = plugin.class.instance_name
+        self.class.send(:attr_reader, name)
+        instance_variable_set("@#{name}", plugin)
+      end
+      # Keep track fo this for `method_missing`
+      @core_plugins = plugin_host.core_plugins
+    end
+
     # @return [String] a string useful to represent the Dangerfile in a message
     #         presented to the user.
     #
     def to_s
       "Dangerfile"
-    end
-
-    # These are the classes that are allowed to also use method_missing
-    # in order to provide broader plugin support
-    def self.core_plugin_classes
-      [DangerfileMessagingPlugin]
-    end
-
-    # The ones that everything would break without
-    def self.essential_plugin_classes
-      [DangerfileMessagingPlugin, DangerfileGitPlugin, DangerfileImportPlugin, DangerfileGitHubPlugin]
     end
 
     # Both of these methods exist on all objects
@@ -52,7 +51,7 @@ module Danger
     end
 
     # When an undefined method is called, we check to see if it's something
-    # that the core DSLs have, then starts looking at plugins support.
+    # that the core DSLs have, then starts looking at plugin support.
 
     # rubocop:disable Style/MethodMissing
 
@@ -63,102 +62,6 @@ module Danger
         end
       end
       super
-    end
-
-    def initialize(env_manager, cork_board)
-      @plugins = {}
-      @core_plugins = []
-      @ui = cork_board
-
-      # Triggers the core plugins
-      @env = env_manager
-
-      # Triggers local plugins from the root of a project
-      Dir["./danger_plugins/*.rb"].each do |file|
-        require File.expand_path(file)
-      end
-
-      refresh_plugins if env_manager.pr?
-    end
-
-    # Iterate through available plugin classes and initialize them with
-    # a reference to this Dangerfile
-    def refresh_plugins
-      plugins = Plugin.all_plugins
-      plugins.each do |klass|
-        next if klass.respond_to?(:singleton_class?) && klass.singleton_class?
-        plugin = klass.new(self)
-        next if plugin.nil? || @plugins[klass]
-
-        name = plugin.class.instance_name
-        self.class.send(:attr_reader, name)
-        instance_variable_set("@#{name}", plugin)
-
-        @plugins[klass] = plugin
-        @core_plugins << plugin if self.class.core_plugin_classes.include? klass
-      end
-    end
-    alias init_plugins refresh_plugins
-
-    def core_dsl_attributes
-      @core_plugins.map { |plugin| { plugin: plugin, methods: plugin.public_methods(false) } }
-    end
-
-    def external_dsl_attributes
-      plugins.values.reject { |plugin| @core_plugins.include? plugin } .map { |plugin| { plugin: plugin, methods: plugin.public_methods(false) } }
-    end
-
-    def method_values_for_plugin_hashes(plugin_hashes)
-      plugin_hashes.flat_map do |plugin_hash|
-        plugin = plugin_hash[:plugin]
-        methods = plugin_hash[:methods].select { |name| plugin.method(name).parameters.empty? }
-
-        methods.map do |method|
-          case method
-          when :api
-            value = "Octokit::Client"
-
-          when :pr_json
-            value = "[Skipped]"
-
-          when :pr_body
-            value = plugin.send(method)
-            value = value.scan(/.{,80}/).to_a.each(&:strip!).join("\n")
-
-          else
-            value = plugin.send(method)
-            # So that we either have one value per row
-            # or we have [] for an empty array
-            value = value.join("\n") if value.kind_of?(Array) && value.count > 0
-          end
-
-          [method.to_s, value]
-        end
-      end
-    end
-
-    # Iterates through the DSL's attributes, and table's the output
-    def print_known_info
-      rows = []
-      rows += method_values_for_plugin_hashes(core_dsl_attributes)
-      rows << ["---", "---"]
-      rows += method_values_for_plugin_hashes(external_dsl_attributes)
-      rows << ["---", "---"]
-      rows << ["SCM", env.scm.class]
-      rows << ["Source", env.ci_source.class]
-      rows << ["Requests", env.request_source.class]
-      rows << ["Base Commit", env.meta_info_for_base]
-      rows << ["Head Commit", env.meta_info_for_head]
-
-      params = {}
-      params[:rows] = rows.each { |current| current[0] = current[0].yellow }
-      params[:title] = "Danger v#{Danger::VERSION}\nDSL Attributes".green
-
-      ui.section("Info:") do
-        ui.puts
-        ui.puts Terminal::Table.new(params)
-        ui.puts
-      end
     end
 
     # Parses the file at a path, optionally takes the content of the file for DI
@@ -200,43 +103,9 @@ module Danger
       end
     end
 
-    def print_results
-      status = status_report
-      return if (status[:errors] + status[:warnings] + status[:messages] + status[:markdowns]).count.zero?
-
-      ui.section("Results:") do
-        [:errors, :warnings, :messages].each do |key|
-          formatted = key.to_s.capitalize + ":"
-          title = case key
-                  when :errors
-                    formatted.red
-                  when :warnings
-                    formatted.yellow
-                  else
-                    formatted
-                  end
-          rows = status[key]
-          print_list(title, rows)
-        end
-
-        if status[:markdowns].count > 0
-          ui.section("Markdown:") do
-            status[:markdowns].each do |current_markdown|
-              ui.puts current_markdown
-            end
-          end
-        end
-      end
-    end
-
-    private
-
-    def print_list(title, rows)
-      ui.title(title) do
-        rows.each do |row|
-          ui.puts("- [ ] #{row}")
-        end
-      end unless rows.empty?
+    # Mainly a helper method for `local --pry`
+    def plugins
+      env.plugin_host.plugins
     end
   end
 end
