+++ lib/danger/plugin_support/plugin.rb
@@ -1,3 +1,5 @@
+require "danger/danger_core/plugin_host"
+
 module Danger
   class Plugin
     def initialize(dangerfile)
@@ -28,7 +30,7 @@ module Danger
     end
 
     def self.clear_external_plugins
-      @all_plugins = @all_plugins.select { |plugin| Dangerfile.essential_plugin_classes.include? plugin }
+      @all_plugins = @all_plugins.select { |plugin| PluginHost.essential_plugin_classes.include? plugin }
     end
 
     def self.inherited(plugin)
