+++ spec/lib/danger/danger_core/plugin_host_spec.rb
@@ -1,0 +1,7 @@
+# it "should add a plugin to the @plugins array" do
+#   class DangerTestPlugin < Danger::Plugin; end
+#   allow(Danger::Plugin).to receive(:all_plugins).and_return([DangerTestPlugin])
+
+#   dm = testing_dangerfile
+#   expect(dm.instance_variable_get("@plugins").length).to eq(1)
+# end
